backend/requirements.txt
fastapi
gunicorn
uvicorn
python-dotenv
pydantic
